+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 1.0.6
+
+- Fixes issue when it was impossible to set the `v-model` value from code
+- Addss button which cleans the value in `dev/MaterialModel.vue`
+- Changed how `npm run test` works
+
 ## Version 1.0.5
 
 - Fixes the error when `v-model` was passed as `null` and no listener was attached